+++ src/assets/locales/es/common.js
@@ -9,7 +9,11 @@ export default {
     phone: 'Teléfono',
     label_message: 'Mensaje: ',
     label_description: 'Descripcion: ',
-    label_preview: 'Vista Previa: '
+    label_preview: 'Vista Previa: ',
+    created_by: 'created by:',
+    canceled_by: 'canceled by:',
+    terminated_by: 'terminated by:',
+    assigned_by: 'assigned by:'
   },
   models: {
     users: 'Usuario | Usuarios',
@@ -28,6 +32,7 @@ export default {
     terminate: 'Terminar',
     filter: 'Filtrar',
     see: 'Ver',
+    copy_phone: 'Copiar teléfono',
     clear_filters: 'Limpiar Filtros',
     items_pages: 'Elementos Por Pagina'
   },
@@ -40,6 +45,8 @@ export default {
     forbidden: 'Permiso denegado!',
     error_chatBot: 'El chatBot y las notificaciones de WhatsApp no pueden estar habilitadas al mismo tiempo!',
     error_assistant: 'El chatBot y el asistente no pueden estar habilitados al mismo tiempo!',
+    copied: 'Copiado al portapapeles!',
+    new_message: 'Nuevo mensaje',
   },
   placeholders: {
     name: 'Ingrese el nombre',
@@ -53,6 +60,8 @@ export default {
     search: 'Buscar',
     map: 'Mapa',
     all: 'Todos',
+    dark: 'Tema Oscuro',
+    light: 'Tema Claro',
     device: 'No tiene dispositivo registrado',
     cancel_percent: 'Índice de cancelación',
     service_info: 'Información del servicio',