+++ src/constants/WhatsappServices.ts
@@ -1,0 +1,5 @@
+export enum WhatsappServices {
+    WHATSAPP_WEB_JS = 'whatsapp-web-js',
+    OFFICIAL = 'api-official',
+    BAILEYS = 'baileys',
+}