+++ src/shims-vue.d.ts
@@ -4,3 +4,8 @@ declare module '*.vue' {
   const component: DefineComponent<{}, {}, any>
   export default component
 }
+
+declare module '*.mp3' {
+  const src: string;
+  export default src;
+}